src/api/requirements.txt
Django==4.0.5
djangorestframework==3.13.1
psycopg2-binary==2.9.9
python-telegram-bot==20.6 
python-dotenv==1.0.0
